+++ pioreactor/actions/__init__.py
@@ -3,9 +3,7 @@ from __future__ import annotations
 
 from pioreactor.actions import led_intensity
 from pioreactor.actions import od_blank
-from pioreactor.actions import od_calibration
 from pioreactor.actions import pump
-from pioreactor.actions import pump_calibration
 from pioreactor.actions import self_test
 from pioreactor.actions.leader import backup_database
 from pioreactor.actions.leader import experiment_profile