+++ pioreactor/cli/calibrations.py
@@ -7,65 +7,19 @@ import click
 from msgspec.yaml import decode as yaml_decode
 from msgspec.yaml import encode as yaml_encode
 
-from pioreactor import structs
 from pioreactor.calibrations.utils import curve_to_callable
 from pioreactor.calibrations.utils import plot_data
+from pioreactor.calibrations import CALIBRATION_PATH
+from pioreactor.calibrations import calibration_assistants
+from pioreactor.calibrations import load_calibration
 from pioreactor.utils import local_persistant_storage
 from pioreactor.whoami import is_testing_env
-
-if not is_testing_env():
-    CALIBRATION_PATH = Path("/home/pioreactor/.pioreactor/storage/calibrations/")
-else:
-    CALIBRATION_PATH = Path(".pioreactor/storage/calibrations/")
-
-# Lookup table for different calibration assistants
-CALIBRATION_ASSISTANTS = {}
-
-
-class CalibrationAssistant:
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        CALIBRATION_ASSISTANTS[cls.target_calibration_type] = cls
-
-    def run(self, *args, **kwargs):
-        raise NotImplementedError("Subclasses must implement this method.")
-
-
-class ODAssistant(CalibrationAssistant):
-    target_calibration_type = "od"
-    calibration_struct = structs.ODCalibration
-
-    def __init__(self):
-        pass
-
-
-class PumpAssistant(CalibrationAssistant):
-    target_calibration_type = "pump"
-    calibration_struct = structs.PumpCalibration
-
-    def __init__(self):
-        pass
-
-
-class StirringAssistant(CalibrationAssistant):
-    target_calibration_type = "stirring"
-    calibration_struct = structs.StirringCalibration
-
-    def __init__(self):
-        pass
-
-    def run(self, min_dc: str | None = None, max_dc: str | None = None) -> structs.StirringCalibration:
-        from pioreactor.calibrations.stirring_calibration import run_stirring_calibration
-
-        return run_stirring_calibration(
-            min_dc=float(min_dc) if min_dc is not None else None, max_dc=float(max_dc) if max_dc else None
-        )
 
 
 @click.group(short_help="calibration utils")
 def calibration():
     """
-    Calibration CLI - A unified interface for all calibration types.
+    interface for all calibration types.
     """
     pass
 
@@ -81,18 +35,18 @@ def list_calibrations(cal_type: str):
         click.echo(f"No calibrations found for type '{cal_type}'. Directory does not exist.")
         return
 
-    assistant = CALIBRATION_ASSISTANTS.get(cal_type)
+    assistant = calibration_assistants.get(cal_type)
 
-    header = f"{'Name':<50}{'Created At':<25}{'Subtype':<15}{'Current?':<15}"
+    header = f"{'Name':<50}{'Created At':<25}{'Subtype':<15}{'Active?':<15}"
     click.echo(header)
     click.echo("-" * len(header))
 
-    with local_persistant_storage("current_calibrations") as c:
+    with local_persistant_storage("active_calibrations") as c:
         for file in calibration_dir.glob("*.yaml"):
             try:
                 data = yaml_decode(file.read_bytes(), type=assistant.calibration_struct)
-                current = c.get((cal_type, data.calibration_subtype)) == data.calibration_name
-                row = f"{data.calibration_name:<50}{data.created_at.strftime('%Y-%m-%d %H:%M:%S'):<25}{data.calibration_subtype or '':<15}{'✅' if current else '':<15}"
+                active = c.get((cal_type, data.calibration_subtype)) == data.calibration_name
+                row = f"{data.calibration_name:<50}{data.created_at.strftime('%Y-%m-%d %H:%M:%S'):<25}{data.calibration_subtype or '':<15}{'✅' if active else '':<15}"
                 click.echo(row)
             except Exception as e:
                 error_message = f"Error reading {file.stem}: {e}"
@@ -109,9 +63,9 @@ def run_calibration(ctx, cal_type: str):
     """
 
     # Dispatch to the assistant function for that type
-    assistant = CALIBRATION_ASSISTANTS.get(cal_type)
+    assistant = calibration_assistants.get(cal_type)
     if assistant is None:
-        click.echo(f"No assistant found for calibration type '{cal_type}'.")
+        click.echo(f"No assistant found for calibration type '{cal_type}'. Available types: {list(calibration_assistants.keys())}")
         raise click.Abort()
 
     # Run the assistant function to get the final calibration data
@@ -130,6 +84,11 @@ def run_calibration(ctx, cal_type: str):
 
     # TODO: send to leader
 
+    # make active
+    with local_persistant_storage("active_calibrations") as c:
+        c[(cal_type, calibration_data.calibration_subtype)] = calibration_name
+
+
     click.echo(f"Calibration '{calibration_name}' of type '{cal_type}' saved to {out_file}")
 
 
@@ -140,17 +99,7 @@ def display_calibration(cal_type: str, calibration_name: str):
     """
     Display the contents of a calibration YAML file.
     """
-    file = CALIBRATION_PATH / cal_type / f"{calibration_name}.yaml"
-    if not file.exists():
-        click.echo(f"No such calibration file: {file}")
-        raise click.Abort()
-
-    assistant = CALIBRATION_ASSISTANTS.get(cal_type)
-
-    try:
-        data = yaml_decode(file.read_bytes(), type=assistant.calibration_struct)
-    except Exception as e:
-        click.echo(f"Error reading {file.stem()}: {e}")
+    data = load_calibration(cal_type, calibration_name)
 
     click.echo()
     curve = curve_to_callable(data.curve_type, data.curve_data_)
@@ -167,32 +116,27 @@ def display_calibration(cal_type: str, calibration_name: str):
     click.echo()
     click.echo("==== YAML output ====")
     click.echo()
-    click.echo(file.read_text())
+    click.echo(yaml_encode(data))
 
 
-@calibration.command(name="set-current")
-@click.option("--type", "cal_type", required=True, help="Which calibration type to set as current.")
-@click.option("--name", "calibration_name", required=True, help="Which calibration name to set as current.")
-def set_current_calibration(cal_type: str, calibration_name: str):
+@calibration.command(name="set-active")
+@click.option("--type", "cal_type", required=True, help="Which calibration type to set as active.")
+@click.option("--name", "calibration_name", required=False, help="Which calibration name to set as active.")
+def set_active_calibration(cal_type: str, calibration_name: str | None):
     """
-    Mark a specific calibration as 'current' for that calibration type.
+    Mark a specific calibration as 'active' for that calibration type.
     """
 
-    file = CALIBRATION_PATH / cal_type / f"{calibration_name}.yaml"
+    if calibration_name is None:
+        click.echo("No calibration name provided. Clearing active calibration.")
+        with local_persistant_storage("active_calibrations") as c:
+            c.pop((cal_type, None))
 
-    # Dispatch to the assistant function for that type
-    assistant = CALIBRATION_ASSISTANTS.get(cal_type)
-    if assistant is None:
-        click.echo(f"No assistant found for calibration type '{cal_type}'.")
-        raise click.Abort()
+    else:
+        data = load_calibration(cal_type, calibration_name)
 
-    try:
-        data = yaml_decode(file.read_bytes(), type=assistant.calibration_struct)
-    except Exception as e:
-        click.echo(f"Error reading {file.stem()}: {e}")
-
-    with local_persistant_storage("current_calibrations") as c:
-        c[(cal_type, data.calibration_subtype)] = calibration_name
+        with local_persistant_storage("active_calibrations") as c:
+            c[(data.calibration_type, data.calibration_subtype)] = data.calibration_name
 
 
 @calibration.command(name="delete")
@@ -213,3 +157,6 @@ def delete_calibration(cal_type: str, calibration_name: str):
 
     target_file.unlink()
     click.echo(f"Deleted calibration '{calibration_name}' of type '{cal_type}'.")
+
+    # TODO: delete from leader and handle updating active?
+